Perform a HEAD before the actual request. Close #340

+++ src/FeedIo/Adapter/FileSystem/Response.php
@@ -33,7 +33,7 @@ class Response implements ResponseInterface
     /**
      * @return int
      */
-    public function getDuration(): int
+    public function getDuration(): float
     {
         return 0;
     }
